refactor: split audioactiondialog and generatesheet under 500 lines
Both components exceeded the project's 500-line guideline. Pure
extraction of presentational JSX and self-contained logic — no behavior
changes:

- AudioActionDialog.tsx (791 -> 515 lines): extracted the MediaRecorder
  start/stop/timer logic into useAudioRecording, and the mode-tabs,
  analyzing-state, upload, and preview/analysis/actions sections into
  their own components under audio-action-dialog/
- GenerateSheet.tsx (667 -> 484 lines): extracted the seven secondary
  dialogs (project/artist/voice-clone/audio-action/lyrics-assistant/
  history/style-preset) into GenerateSheetDialogs
