+++ docs/OPTIMIZATION_PLAN.md
@@ -26,6 +26,8 @@
 | Time to Interactive    | 4.5s    | <3.5s   | <2.8s   |
 | TypeScript coverage    | 85%     | 95%     | 100%    |
 
+> **Апдейт 2026-07-03**: "Bundle size" здесь исторически измерялся как сумма всех JS-чанков (`size-limit`'s "Total Bundle") — сейчас это 2.11MB gzip, т.к. метрика включает admin/studio/lazy-страницы, которые обычный пользователь не грузит. Более показательная метрика — eager JS на холодной загрузке любой страницы: снижена с ~1.19MB до ~508KB gzip (см. `docs/BUNDLE_ANALYSIS.md`, раздел "Update 2026-07-03"). TypeScript `any`-бюджет уже закрыт (0/50, было 342 на момент последнего аудита) — эта строка KPI таблицы ниже устарела, если такая есть.
+
 ---
 
 ## 🏗️ Архитектура оптимизаций
