+++ src/components/generate-form/audio-action-dialog/useAudioRecording.ts
@@ -1,0 +1,86 @@
+import { useState, useRef, useEffect } from "react";
+import { toast } from "sonner";
+import { logger } from "@/lib/logger";
+
+interface UseAudioRecordingOptions {
+  /** Called with the recorded file + object URL once recording stops */
+  onRecorded: (file: File, url: string) => void;
+}
+
+/** MediaRecorder capture (start/stop/timer) used by AudioActionDialog's "record" mode. */
+export function useAudioRecording({ onRecorded }: UseAudioRecordingOptions) {
+  const [isRecording, setIsRecording] = useState(false);
+  const [recordingTime, setRecordingTime] = useState(0);
+
+  const mediaRecorderRef = useRef<MediaRecorder | null>(null);
+  const chunksRef = useRef<Blob[]>([]);
+  const recordingIntervalRef = useRef<NodeJS.Timeout | null>(null);
+
+  useEffect(() => {
+    if (isRecording) {
+      recordingIntervalRef.current = setInterval(() => {
+        setRecordingTime((prev) => prev + 1);
+      }, 1000);
+    } else {
+      if (recordingIntervalRef.current) {
+        clearInterval(recordingIntervalRef.current);
+      }
+      setRecordingTime(0);
+    }
+    return () => {
+      if (recordingIntervalRef.current) {
+        clearInterval(recordingIntervalRef.current);
+      }
+    };
+  }, [isRecording]);
+
+  const startRecording = async () => {
+    try {
+      const stream = await navigator.mediaDevices.getUserMedia({
+        audio: {
+          echoCancellation: true,
+          noiseSuppression: true,
+          sampleRate: 44100,
+        },
+      });
+
+      const mimeType = MediaRecorder.isTypeSupported("audio/webm;codecs=opus") ? "audio/webm;codecs=opus" : "audio/mp4";
+
+      const mediaRecorder = new MediaRecorder(stream, { mimeType });
+      mediaRecorderRef.current = mediaRecorder;
+      chunksRef.current = [];
+
+      mediaRecorder.ondataavailable = (e) => {
+        if (e.data.size > 0) {
+          chunksRef.current.push(e.data);
+        }
+      };
+
+      mediaRecorder.onstop = () => {
+        const blob = new Blob(chunksRef.current, { type: mimeType });
+        const ext = mimeType.includes("webm") ? "webm" : "mp4";
+        const file = new File([blob], `recording-${Date.now()}.${ext}`, { type: mimeType });
+        const url = URL.createObjectURL(blob);
+
+        stream.getTracks().forEach((track) => track.stop());
+        onRecorded(file, url);
+      };
+
+      mediaRecorder.start(1000);
+      setIsRecording(true);
+      toast.success("Запись началась");
+    } catch (error) {
+      logger.error("Recording error", { error });
+      toast.error("Не удалось получить доступ к микрофону");
+    }
+  };
+
+  const stopRecording = () => {
+    if (mediaRecorderRef.current && isRecording) {
+      mediaRecorderRef.current.stop();
+      setIsRecording(false);
+    }
+  };
+
+  return { isRecording, recordingTime, startRecording, stopRecording };
+}
