+++ src/components/home/DiscoverTabs.tsx
@@ -8,7 +8,8 @@
  * Goal: drastically reduce visual noise on the home page.
  */
 
-import { memo, useState } from "react";
+import { memo, useState, forwardRef, useMemo, useCallback } from "react";
+import { VirtuosoGrid } from "react-virtuoso";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { UnifiedTrackCard } from "@/components/track/track-card-new";
 import { Button } from "@/components/ui/button";
@@ -17,6 +18,7 @@ import { cn } from "@/lib/utils";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { motion, type Variants } from "@/lib/motion";
 import { useReducedMotion } from "@/hooks/useReducedMotion";
+import { GridSkeleton } from "@/components/ui/skeleton-components";
 import type { TrackData } from "@/components/track/track-card-new/types";
 
 const gridVariants: Variants = {
@@ -28,6 +30,15 @@ const cardVariants: Variants = {
   hidden: { opacity: 0, y: 18, scale: 0.96 },
   visible: { opacity: 1, y: 0, scale: 1, transition: { duration: 0.32, ease: [0.16, 1, 0.3, 1] } },
 };
+
+const fadeInVariants: Variants = {
+  hidden: { opacity: 0 },
+  visible: { opacity: 1, transition: { duration: 0.2 } },
+};
+
+// Only worth windowing once a list has grown past several "load more" pages —
+// the homepage's first page (pageSize=20) always uses the plain animated Grid.
+const VIRTUALIZE_THRESHOLD = 24;
 
 interface DiscoverTabsProps {
   popularTracks: TrackData[];
@@ -92,6 +103,78 @@ const Grid = memo(function Grid({
   );
 });
 
+// Windowed variant for long, paginated lists (>= VIRTUALIZE_THRESHOLD). Per-item
+// hover/tap motion is kept, but the batch stagger-entrance animation is dropped
+// in favor of a plain fade-in — staggerChildren doesn't make sense once items
+// are recycled in and out of the DOM by the virtualizer.
+const VirtualizedGrid = memo(function VirtualizedGrid({
+  tracks,
+  columns,
+  onTrackClick,
+  onRemix,
+}: {
+  tracks: TrackData[];
+  columns: number;
+  onTrackClick?: (track: TrackData) => void;
+  onRemix?: (id: string) => void;
+}) {
+  const prefersReducedMotion = useReducedMotion();
+
+  const GridContainer = useMemo(
+    () =>
+      forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(({ children, style, ...props }, ref) => (
+        <div
+          ref={ref}
+          style={{ ...style, display: "grid", gridTemplateColumns: `repeat(${columns}, minmax(0, 1fr))` }}
+          className="gap-3 sm:gap-4"
+          {...props}
+        >
+          {children}
+        </div>
+      )),
+    [columns],
+  );
+
+  const gridComponents = useMemo(() => ({ List: GridContainer }), [GridContainer]);
+
+  const computeItemKey = useCallback((index: number) => tracks[index]?.id ?? `track-${index}`, [tracks]);
+
+  const renderItem = useCallback(
+    (index: number) => {
+      const track = tracks[index];
+      if (!track) return null;
+      return (
+        <motion.div
+          initial={prefersReducedMotion ? undefined : "hidden"}
+          animate={prefersReducedMotion ? undefined : "visible"}
+          variants={prefersReducedMotion ? undefined : fadeInVariants}
+          whileHover={prefersReducedMotion ? undefined : { y: -3 }}
+          whileTap={prefersReducedMotion ? undefined : { scale: 0.97 }}
+        >
+          <UnifiedTrackCard
+            track={track}
+            variant="grid"
+            onPlay={() => onTrackClick?.(track)}
+            showActions={false}
+            data-onremix={onRemix ? "true" : undefined}
+          />
+        </motion.div>
+      );
+    },
+    [tracks, onTrackClick, onRemix, prefersReducedMotion],
+  );
+
+  return (
+    <VirtuosoGrid
+      useWindowScroll
+      totalCount={tracks.length}
+      computeItemKey={computeItemKey}
+      components={gridComponents}
+      itemContent={renderItem}
+    />
+  );
+});
+
 const LoadMore = ({ visible, loading, onClick }: { visible?: boolean; loading?: boolean; onClick?: () => void }) => {
   if (!visible || !onClick) return null;
   return (
@@ -108,6 +191,20 @@ const LoadMore = ({ visible, loading, onClick }: { visible?: boolean; loading?: 
     </div>
   );
 };
+
+// Picks the plain animated grid for short lists and only switches to windowed
+// rendering once a list has grown past several "load more" pages.
+function TracksGrid(props: {
+  tracks: TrackData[];
+  columns: number;
+  onTrackClick?: (track: TrackData) => void;
+  onRemix?: (id: string) => void;
+}) {
+  if (!props.tracks.length) {
+    return <p className="text-sm text-muted-foreground text-center py-8">Пока ничего нет</p>;
+  }
+  return props.tracks.length >= VIRTUALIZE_THRESHOLD ? <VirtualizedGrid {...props} /> : <Grid {...props} />;
+}
 
 export const DiscoverTabs = memo(function DiscoverTabs({
   popularTracks,
@@ -141,10 +238,10 @@ export const DiscoverTabs = memo(function DiscoverTabs({
 
       <TabsContent value="popular" className="mt-0">
         {isLoading && popularTracks.length === 0 ? (
-          <SkeletonGrid columns={columns} />
+          <GridSkeleton columns={columns} count={columns * 2} />
         ) : (
           <>
-            <Grid tracks={popularTracks} columns={columns} onTrackClick={onTrackClick} onRemix={onRemix} />
+            <TracksGrid tracks={popularTracks} columns={columns} onTrackClick={onTrackClick} onRemix={onRemix} />
             <LoadMore visible={hasMorePopular} loading={isLoadingMorePopular} onClick={onLoadMorePopular} />
           </>
         )}
@@ -152,10 +249,10 @@ export const DiscoverTabs = memo(function DiscoverTabs({
 
       <TabsContent value="new" className="mt-0">
         {isLoading && recentTracks.length === 0 ? (
-          <SkeletonGrid columns={columns} />
+          <GridSkeleton columns={columns} count={columns * 2} />
         ) : (
           <>
-            <Grid tracks={recentTracks} columns={columns} onTrackClick={onTrackClick} onRemix={onRemix} />
+            <TracksGrid tracks={recentTracks} columns={columns} onTrackClick={onTrackClick} onRemix={onRemix} />
             <LoadMore visible={hasMoreRecent} loading={isLoadingMoreRecent} onClick={onLoadMoreRecent} />
           </>
         )}
@@ -163,14 +260,6 @@ export const DiscoverTabs = memo(function DiscoverTabs({
     </Tabs>
   );
 });
-
-const SkeletonGrid = ({ columns }: { columns: number }) => (
-  <div className="grid gap-3" style={{ gridTemplateColumns: `repeat(${columns}, minmax(0, 1fr))` }}>
-    {Array.from({ length: columns * 2 }).map((_, i) => (
-      <div key={i} className="aspect-square rounded-xl bg-muted/20 animate-pulse" />
-    ))}
-  </div>
-);
 
 // Suppress unused-import warning for Clock (kept for future "Recent" tab)
 void Clock;
